% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calib_helpers.R
\name{Corner_text}
\alias{Corner_text}
\title{Add text to the corner of a plot}
\usage{
Corner_text(text, location = "topleft")
}
\arguments{
\item{text}{text to be added to the plot as a language object}

\item{location}{location within the plot for the text to be placed}
}
\value{
data
}
\description{
Add text to a location within a plot. <https://github.com/aemon-j/gotmtools/blob/master/R/Corner_text.R>
}
