\name{depthsampleGLM}
\alias{depthsampleGLM}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{gets wtr temperature observations at depths from GLM data.frame}
\description{This function gets wtr temperature observations at depths from GLM data.frame}
\usage{
depthsampleGLM(GLM, sampleDepths)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{GLM}{GLM data.frame}
  \item{sampleDepths}{depths for subsampling}
}
\value{
GLM data.frame
}
\author{
Jordan S. Read & Luke A. Winslow
}

\keyword{rGLM}
