% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot_compare_to_field.R
\name{.compare_to_field}
\alias{.compare_to_field}
\title{Internal field data compare function.}
\usage{
.compare_to_field(compare_data, bthA, bthD, metric, as_value = FALSE,
  na.rm = TRUE)
}
\arguments{
\item{compare_data}{A data.frame with columns "DateTime, "Depth", "Obs" and "Mod", in that order.}

\item{bthA}{a numeric vector of cross sectional areas (m^2) corresponding to bthD depths}

\item{bthD}{a numeric vector of depths (m) which correspond to areal measures in bthA}

\item{metric}{a string representing a physical metric. 
Should be a rLakeAnalyzer function or other valid function.}

\item{as_value}{a boolean for calculating RMSE (F) or returning all values (T)}

\item{na.rm}{a boolean for remove NAs (only used if as_values == F)}
}
\value{
See return options for \link{compare_to_field}
}
\description{
Used in \link{compare_to_field}. Accepts data.frame instead 
of nc and field file locations
}
