% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nml_helpers.R
\name{from.glm_boolean}
\alias{from.glm_boolean}
\title{go from glm2.nml logical vectors to R logicals}
\usage{
from.glm_boolean(values)
}
\arguments{
\item{values}{a vector of strings containing either .false. or .true.}
}
\value{
a logical vector
}
\description{
go from glm2.nml logical vectors to R logicals
}
\keyword{internal}
