\name{get.bth}
\alias{get.bth}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{gets bathymetry file from nml file}
\description{This function hypsography data for .bth file from GLM nml.}
\usage{
get.bth(nml)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{nml}{an R nml list}
}
\value{
a data.frame with hypsography observations
}
\author{
Jordan S. Read
}

\keyword{rGLM}
