\name{get.lvl}
\alias{get.lvl}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{gets time series level data for .lvl file}
\description{This function gets time series level data for .lvl file from GLM nc output.}
\usage{
get.lvl(GLMnc,nml)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{GLMnc}{a GLMnc object}
  \item{nml}{an R nml list}
}
\value{
a timeseries data.frame with level observations
}
\author{
Jordan S. Read
}

\keyword{rGLM}
