\name{get.nml}
\alias{get.nml}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{gets value in R nml list}
\description{This function gets the specified value in the nml list for GLM.}
\usage{
get.nml(nml,argName)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{nml}{ an nml list}
  \item{argName}{ a valid nml parameter}
}
\value{
value of specified nml parameter
}
\author{
Jordan S. Read
}

\keyword{rGLM}
