\name{get.wnd}
\alias{get.wnd}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{gets time series .wnd file from GLM nc output}
\description{This function gets time series .wnd file from GLM nc output}
\usage{
get.wtr(GLMnc)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{GLMnc}{GLM nc object}
}
\value{
a data.frame time series wind speed observations
}
\author{
Jordan S. Read
}

\keyword{rGLM}
