% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_calib_init_validation.R
\name{get_calib_init_validation}
\alias{get_calib_init_validation}
\title{Get last values from a calibration period to be used for the validation}
\usage{
get_calib_init_validation(nml_file, output)
}
\arguments{
\item{nml_file}{String of the GLM namelist file, in most cases this is 'glm3.nml'}

\item{output}{String of the file path in which the output.nc file is stored (created by calibration period)}
}
\description{
Reads in nml-file and output from calibration to get initial values for the calibration
}
\examples{
nc_file <- system.file("extdata", "output/output.nc", package = "glmtools")
nml_file <- system.file("extdata", "glm3.nml", package = "glmtools")
initvalues <- get_calib_init_validation(nml = nml_file, output = nc_file)
}
\seealso{
\code{\link{get_calib_setup}, \link{get_calib_periods}, \link{calibrate_sim}}
}
\author{
Robert Ladwig
}
\keyword{methods}
