% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_evaporation.R
\name{get_evaporation}
\alias{get_evaporation}
\title{get evaporation from GLM simulation}
\usage{
get_evaporation(file)
}
\arguments{
\item{file}{a string with the path to the netcdf output from GLM}
}
\value{
a data.frame with DateTime and evaporation (in mm/day)
}
\description{
Creates a data.frame with DateTime and E (mm/day).  \cr
}
\examples{
sim_folder <- run_example_sim(verbose = FALSE)
nc_file <- file.path(sim_folder, 'output.nc')
evap <- get_evaporation(nc_file)
plot(evap)
}
\author{
Jordan S. Read
}
\keyword{methods}

