% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get_ice}
\alias{get_ice}
\title{get ice depth from GLM simulation}
\usage{
get_ice(file)
}
\arguments{
\item{file}{a string with the path to the netcdf output from GLM}
}
\value{
a data.frame with DateTime and ice
}
\description{
Creates a data.frame with DateTime and ice.  \cr
}
\examples{
file = '../test/output.nc'
ice <- get_ice(file)
}
\author{
Luke A. Winslow, Jordan S. Read
}
\keyword{methods}

