% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_nml_value}
\alias{get_nml_value}
\title{gets a nml value according to an arg_name}
\usage{
get_nml_value(glm_nml, arg_name)
}
\arguments{
\item{glm_nml}{a nml (a list) for GLM config}

\item{arg_name}{a string representing a valid field in glm_nml}
}
\value{
arg_val value for the valid field in glm_nml specified by \code{arg_name}
}
\description{
This function returns an nml value according to the arg_name nml list for GLM.
}
\examples{
nml_file <- system.file('extdata', 'glm.nml', package = 'glmtools')
glm_nml <- read_nml(nml_file)
get_nml_value(glm_nml,arg_name = 'Kw')
}
\author{
Jordan S. Read
}
\seealso{
\link{read_nml}, \link{set_nml}
}

