% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_surface_height.R
\name{get_surface_height}
\alias{get_surface_height}
\title{get surface height from GLM simulation}
\usage{
get_surface_height(file = "output.nc", ice.rm = TRUE, snow.rm = TRUE, ...)
}
\arguments{
\item{file}{a string with the path to the netcdf output from GLM}

\item{ice.rm}{a boolean for including ice thickness in surface height}

\item{snow.rm}{a boolean for including snow depth thickness in surface height}

\item{...}{additional arguments passed to \code{\link{resample_sim}}}
}
\value{
a data.frame with DateTime and surface_height (in meters)
}
\description{
Creates a data.frame with DateTime and surface_height.  \cr
}
\examples{
sim_folder <- run_example_sim(verbose = FALSE)
nc_file <- file.path(sim_folder, 'output.nc')
surface <- get_surface_height(file = nc_file)
surface_w_ice <- get_surface_height(file = nc_file, ice.rm = FALSE, snow.rm = FALSE)
}
\author{
Jordan S. Read, Luke A. Winslow
}
\keyword{methods}

