% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_temp.R
\name{get_temp}
\alias{get_temp}
\title{get water temperatures from a GLM simulation}
\usage{
get_temp(
  file = "output.nc",
  reference = "bottom",
  z_out = NULL,
  t_out = NULL,
  ...
)
}
\arguments{
\item{file}{a string with the path to the netcdf output from GLM}

\item{reference}{a string which specifies the vertical reference ('surface' or 'bottom')}

\item{z_out}{an optional vector of depths for temperature output (in meters). 
If NULL, depths will be determined based on the depth of the lake}

\item{t_out}{a vector of POSIXct dates for temporal resampling (order is important)}

\item{...}{additional arguments passed to \code{resample_sim()}}
}
\value{
a data.frame with DateTime and temperature at depth
}
\description{
Creates a data.frame with DateTime and temperatures (in deg C).  \cr
Temperatures that are sampled out of the GLM output are taken relative 
to the surface (\code{reference = 'surface'}) or the bottom of the lake 
(\code{reference = 'bottom'}).
}
\examples{
sim_folder <- run_example_sim(verbose = FALSE)
nc_file <- file.path(sim_folder, 'output/output.nc')
temp_surf <- get_temp(nc_file, reference = 'surface', z_out = c(0,1,2))
temp_bot <- get_temp(nc_file, reference = 'bottom', z_out = c(0,1,2))
temp_bot <- get_temp(nc_file)

#-- get temporal subset--
t_out <- seq(as.POSIXct("2015-04-15"), as.POSIXct("2015-06-01"), by = 86400)
temp_surf <- get_temp(nc_file, reference = 'surface', z_out = 0, t_out = t_out)
plot(temp_surf)
}
\seealso{
\link{resample_sim}
}
\author{
Jordan S. Read, Luke A. Winslow
}
\keyword{methods}
