% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_var.R
\name{get_var}
\alias{get_var}
\title{get variable from a GLM simulation}
\usage{
get_var(
  file = "output.nc",
  var_name,
  reference = "bottom",
  z_out = NULL,
  t_out = NULL,
  ...
)
}
\arguments{
\item{file}{a string with the path to the netcdf output from GLM}

\item{var_name}{a name of a valid variable in the netcdf file specified with \code{file}.}

\item{reference}{a string which specifies the vertical reference ('surface' or 'bottom'). 
(Only used for variables with multiple depths)}

\item{z_out}{an optional vector of depths for value output (in meters). 
(Only used for variables with multiple depths). 
If NULL, depths will be determined based on the depth of the lake}

\item{t_out}{a vector of POSIXct dates for temporal resampling (order is important)}

\item{...}{additional arguments passed to \code{resample_sim()}}
}
\value{
a data.frame with DateTime and variable at depth
}
\description{
Creates a data.frame with DateTime and variable  \cr
Variable is sampled out of the GLM output and is taken relative 
to the surface (\code{reference = 'surface'}) or the bottom of the lake 
(\code{reference = 'bottom'}).
}
\examples{
sim_folder <- run_example_sim(verbose = FALSE)
nc_file <- file.path(sim_folder, 'output.nc')
# list variables in the netcdf output from GLM:
print(sim_vars(nc_file))
evaporation <- get_var(nc_file, var_name = "evap")
plot(evaporation)
}
\seealso{
\code{\link{get_temp}}, \code{\link{plot_var}}
}
\author{
Jordan S. Read, Luke A. Winslow
}
\keyword{methods}
