% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_vars}
\alias{get_vars}
\title{get list of variables from a GLM simulation}
\usage{
get_vars(file)
}
\arguments{
\item{file}{a string with the path to the netcdf output from GLM}
}
\description{
get list of variables from a GLM simulation
}
\examples{
get_vars(file = fabm_sim_nc)
}
\author{
Jordan S. Read, Luke A. Winslow, Stuart E. Jones
}
\keyword{methods}

