\name{getElevGLM}
\alias{getElevGLM}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{gets elevation of observations as a time series from a GLM data.frame}
\description{This function gets elevation of observations as a time series from a GLM data.frame}
\usage{
getElevGLM(GLM)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{GLM}{GLM data.frame}
}
\value{
time series of model layer elevations as a data.frame.
}
\author{
Jordan S. Read
}

\keyword{rGLM}
