\name{getGLMnc}
\alias{getGLMnc}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{opens netCDF file and creates GLMnc object}
\description{This function opens netCDF file and creates GLMnc object}
\usage{
getGLMnc(fileName,folder)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{fileName}{name of GLM output file}
  \item{folder}{location of GLM output file}
}
\value{
GLMnc, a netCDF object for GLM output
}
\author{
Jordan S. Read
}

\keyword{rGLM}
