\name{getIceGLMnc}
\alias{getIceGLMnc}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{gets ice data from GLMnc netCDF object}
\description{This function gets ice data from GLMnc netCDF object}
\usage{
getIceGLMnc(GLMnc)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{GLMnc}{GLM netCDF object}
}
\value{
Ice values for GLM output
}
\author{
Luke A. Winslow & Jordan S. Read
}

\keyword{rGLM}
