\name{getSurfaceElevGLM}
\alias{getSurfaceElevGLM}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{gets elevation of the surface as a time series from a GLM data.frame}
\description{This function gets elevation of the surface as a time series from a GLM data.frame}
\usage{
getSurfaceElevGLM(GLM)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{GLM}{GLM data.frame}
}
\value{
time series of surface elevations as a data.frame.
}
\author{
Jordan S. Read
}

\keyword{rGLM}
