\name{getTimeGLMnc}
\alias{getTimeGLMnc}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{gets time data from GLMnc netCDF object}
\description{This function gets time data from GLMnc netCDF object}
\usage{
getTimeGLMnc(GLMnc)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{GLMnc}{GLM netCDF object}
}
\value{
Time values for GLM output
}
\author{
Luke A. Winslow
}

\keyword{rGLM}
