\name{getTimeInfo}
\alias{getTimeInfo}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{gets time information from GLM netCDF object}
\description{This function gets time information from GLM netCDF object}
\usage{
getTimeInfo(GLMnc)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{GLMnc}{GLM netCDF object}
}
\value{
Time information, start and stop times, time step.
}
\author{
Luke A. Winslow & Jordan S. Read
}

\keyword{rGLM}
