\name{getWndGLMnc }
\alias{getWndGLMnc }
%- Also NEED an '\alias' for EACH other topic documented here.
\title{gets wind data from GLMnc netCDF object}
\description{This function gets wind data from GLMnc netCDF object}
\usage{
getWndGLMnc (GLMnc)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{GLMnc}{GLM netCDF object}
}
\value{
Wind values for GLM output
}
\author{
Luke A. Winslow & Jordan S. Read
}

\keyword{rGLM}
