\name{init.lke}
\alias{init.lke}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{initializes Lake Analyzer list for .lke file}
\description{This function initializes Lake Analyzer list for .lke file.}
\usage{
init.lke()
}
%- maybe also 'usage' for other objects documented here.
\arguments{
}
\value{
list of lke values
}
\author{
Jordan S. Read
}

\keyword{rGLM}
