\name{init.lkeMeta}
\alias{init.lkeMeta}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{initializes Lake Analyzer list for .lke file metadata}
\description{This function initializes Lake Analyzer list for .lke file metadata.}
\usage{
init.lkeMeta()
}
%- maybe also 'usage' for other objects documented here.
\arguments{
}
\value{
list of lke metadata values
}
\author{
Jordan S. Read
}

\keyword{rGLM}
