% Generated by roxygen2 (4.0.2): do not edit by hand
\name{init_lke}
\alias{init_lke}
\title{initializes Lake Analyzer list for .lke file}
\usage{
init_lke()
}
\value{
lke a list with Lake Analyzer parameter call defaults
}
\description{
This function initializes Lake Analyzer list for .lke file.  \cr
}
\examples{
lke <- init_lke()
lke <- set_lke(lke, arg_name='totalDep', arg_val=19.4)
}
\author{
Jordan S. Read
}
\seealso{
\link{write_lke}, \link{set_lke}
}
\keyword{methods}

