% Generated by roxygen2 (4.0.1): do not edit by hand
\name{model_diagnostics}
\alias{model_diagnostics}
\title{run diagnostics on model results}
\usage{
model_diagnostics(nc_file, field_file, metrics, fig_path, report = F, ...)
}
\arguments{
\item{nc_file}{a string with the path to the netcdf output from GLM}

\item{field_file}{a string with the path to the field observation file}

\item{metrics}{a string or vector of strings representing physical metrics.
Should be a rLakeAnalyzer function or other valid function.}

\item{fig_path}{FALSE if plot to screen, string path if save plot as .png.
If argument is not used, plotting is skipped}

\item{report}{default as FALSE, so no stats are returned. If TRUE, return
summary stats for each metric used.}

\item{...}{additional arguments passed to compare_to_field (such as na.rm or nml_file)}
}
\value{
a report or figures comparing modeled results to observations
}
\description{
run diagnostics on model results
}
\examples{
nc_file <- system.file('extdata', 'output.nc', package = 'glmtools')
nml_file <- system.file('extdata', 'glm.nml', package = 'glmtools')
field_file <- system.file('extdata', 'field_data.tsv', package = 'glmtools')

model_diagnostics(nc_file, field_file,
                          metrics = 'center.buoyancy', fig_path = FALSE)
model_diagnostics(nc_file, field_file, nml_file = nml_file,
                          metrics = c('center.buoyancy','thermo.depth', 'schmidt.stability'),
                          fig_path = FALSE)

# write the fig out to file:
model_diagnostics(nc_file, field_file, nml_file = nml_file, fig_path = '../test_fig.png',
                          metrics = c('center.buoyancy','thermo.depth', 'schmidt.stability'))

# return the diagnostic info but avoid the plot:
model_diagnostics(nc_file, field_file, nml_file = nml_file, report = TRUE,
                          metrics = c('center.buoyancy','thermo.depth', 'schmidt.stability'))
}
\author{
Jordan S. Read
}
\seealso{
\link{compare_to_field}, \link{resample_to_field}, \link{read_nml}, \link{get_metrics}
}
\keyword{methods}

