% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_meteo.R
\name{plot_meteo}
\alias{plot_meteo}
\title{Plot meterological drivers from a csv file}
\usage{
plot_meteo(met_file, xmin = NA, xmax = NA, fig_path = NULL, ...)
}
\arguments{
\item{met_file}{Options: 1) file path to glm2.nml file. Date limits pulled from .nml file
2) file path to .csv file. csv file in wide format. First column is Date, additional columns are meteorological paramters. 
3) data.frame object in wide format. First column is Date, additional columns are meteorological paramters.}

\item{xmin}{Optional, start date for xaxis. Formatted as same date format as meteo file.}

\item{xmax}{Optional, end date for xaxis. Formatted as same date format as meteo file.}

\item{fig_path}{Default is NULL (only plots to screen). Enter string path to save as output file. File type can be anything supported by \code{\link[ggplot2:ggsave]{ggplot2:ggsave}}. See examples.}

\item{...}{additional arguments passed to \code{\link[ggplot2:ggsave]{ggplot2:ggsave}}}
}
\description{
Plot meterological drivers from a csv file
}
\examples{
met_file <- system.file("extdata", "LakeMendota_NLDAS.csv", package = "glmtools")
plot_meteo(met_file)
}
\seealso{
\code{\link{read_nml}}
}
\author{
Jordan S. Read, Luke A. Winslow
}
\keyword{methods}
