% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot_temp}
\alias{plot_temp}
\title{plot water temperatures from a GLM simulation}
\usage{
plot_temp(file)
}
\arguments{
\item{file}{a string with the path to the netcdf output from GLM}
}
\description{
plot water temperatures from a GLM simulation
}
\examples{
file <- system.file('extdata', 'output.nc', package = 'glmtools')
plot_temp(file = file)
}
\author{
Jordan S. Read, Luke A. Winslow
}
\seealso{
\link{get_temp}
}
\keyword{methods}

