% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot_temp_compare}
\alias{plot_temp_compare}
\title{Plot matching heatmaps for modeled and observed temp}
\usage{
plot_temp_compare(nc_file, field_file, ...)
}
\arguments{
\item{nc_file}{Netcdf model output file}

\item{field_file}{CSV or TSV field data file (see \link{resample_to_field} for format)}
}
\description{
Plot matching heatmaps for modeled and observed temp
}
\examples{
\dontrun{
nc_file <- system.file('extdata', 'output.nc', package = 'glmtools')
field_file <- system.file('extdata', 'field_data.tsv', package = 'glmtools')

plot_temp_compare(nc_file, field_file) ##makes a plot!
}
}
\author{
Luke Winslow
}
\seealso{
Internally uses \link{get_temp} and \link{resample_to_field}
}

