% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot_validate_profiles}
\alias{plot_validate_profiles}
\title{Plot validation and model temperature profiles}
\usage{
plot_validate_profiles(nc_file, field_file, fig_path = FALSE, ...)
}
\arguments{
\item{nc_file}{a string with the path to the netcdf output from GLM}

\item{field_file}{a string with the path to the field observation file}

\item{fig_path}{FALSE if plot to screen, string path if save plot as .png.
If argument is not used, plotting is skipped}

\item{...}{optional arguments passed to \code{resample_to_field}}
}
\description{
Plot validation and model temperature profiles
}
\examples{
nc_file <- system.file('extdata', 'output.nc', package = 'glmtools')
nml_file <- system.file('extdata', 'glm.nml', package = 'glmtools')
field_file <- system.file('extdata', 'field_data.tsv', package = 'glmtools')

#  create a multiple metric diagnostic fig within R:
plot_validate_profiles(nc_file, field_file, fig_path = FALSE, method = 'interp')
}
\author{
Luke A. Winslow, Jordan S. Read
}
\seealso{
\link{validate_sim}, \link{resample_to_field}
}
\keyword{methods}

