% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_var.R
\name{plot_var}
\alias{plot_var}
\title{Plot variables from a GLM simulation}
\usage{
plot_var(file = "output.nc", var_name, fig_path = FALSE,
  reference = "surface", legend.title = NULL, interval = 0.5,
  text.size = 12, show.legend = TRUE, legend.position = "right",
  plot.title = NULL, ...)
}
\arguments{
\item{file}{a string with the path to the netcdf output from GLM}

\item{var_name}{a character vector of valid variable names (see \code{\link{sim_vars}})}

\item{fig_path}{Logical; F if plot to screen, string path if save plot as .png}

\item{reference}{String; 'surface' or 'bottom. Only used for heatmap plots.}

\item{legend.title}{Vector string; Default (`NULL`) will use variable and units from netcdf file}

\item{interval}{Positive number indicating the depth interval in meters to interpolate output data. Must be less than max depth of lake. Default = 0.5 m.}

\item{text.size}{Integer; Default is 12. Higher values will increase text size in plot.}

\item{show.legend}{Logical; TRUE to show legend (default), FALSE to hide legend}

\item{legend.position}{String; Legend position. Default is 'right'. Options: 'left','right','top','bottom'}

\item{plot.title}{Vector string; Default is no title.}

\item{...}{additional arguments passed to \code{ggsave()}}
}
\description{
Plot variables from a GLM simulation
}
\note{
\code{plot_var} uses the \code{\link[graphics]{layout}} function and so is restricted to a full page display.
When creating a heatmap, the output produced by \code{plot_var} is actually a combination of two plots; 
one is a \code{\link[graphics]{.filled.contour}} plot and the other is a legend.
}
\examples{
sim_folder <- run_example_sim(verbose = FALSE)
nc_file <- file.path(sim_folder, 'output.nc')
vars <- sim_vars(file = nc_file)
plot_var(nc_file, 'u_mean')
#Plotting two variables
plot_var(nc_file, var_name = c('temp','wind'), show.legend = F, text.size = 14, plot.title = c('My Lake: Temp','My Lake: Wind'))
#Saving plot
plot_var(file,var_name = c('temp', 'OXY_oxy'),fig_path = '~/figtest.png', width = 6, height = 2, units = 'in')

\dontrun{
# need to specify a valid .nc file here: 
plot_var(file = fabm_sim_nc.nc,
var_name = 'aed_oxygen_oxy', 
fig_path = 'aed_out.png')
}
}
\seealso{
\code{\link{get_temp}}, \code{\link{sim_var_longname}}, 
\code{\link{sim_vars}}, \code{\link{plot_temp}},  \code{\link{get_var}}
}
\author{
Jordan S. Read, Luke A. Winslow, Hilary A. Dugan
}
\keyword{methods}
