% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_var_compare.R
\name{plot_var_compare}
\alias{plot_var_compare}
\title{Plot matching heatmaps for modeled and observed temp}
\usage{
plot_var_compare(nc_file, field_file, var_name, fig_path = FALSE,
  resample = TRUE, col_lim, ...)
}
\arguments{
\item{nc_file}{Netcdf model output file}

\item{field_file}{CSV or TSV field data file (see \link{resample_to_field} for format)}

\item{var_name}{a character vector of valid variable names (see \code{\link{sim_vars}})}

\item{fig_path}{F if plot to screen, string path if save plot as .png}

\item{resample}{sample the model output to the same time points as the observations?}

\item{col_lim}{range for heatmap (in units of the variable)}

\item{\dots}{additional arguments passed to \code{\link{resample_to_field}}}
}
\description{
Plot matching heatmaps for modeled and observed temp
}
\examples{
sim_folder <- run_example_sim(verbose = FALSE)
nc_file <- file.path(sim_folder, 'output.nc')
nml_file <- file.path(sim_folder, 'glm2.nml')
field_file <- file.path(sim_folder, 'field_data.tsv')

run_glm(sim_folder)

plot_var_compare(nc_file, field_file, 'temp', resample=FALSE) ##makes a plot!

}
\seealso{
Internally uses \link{get_var} and \link{resample_to_field}
}
