\name{pretty.nml}
\alias{pretty.nml}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{prints R nml list with pretty formatting}
\description{This function prints the specified values from the nml list for GLM.}
\usage{
	pretty.nml(nml)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{nml}{ an nml list}
}

\author{
Jordan S. Read
}

\keyword{rGLM}
