% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pretty_nml}
\alias{pretty_nml}
\title{prints R nml list with pretty formatting}
\usage{
pretty_nml(nml)
}
\arguments{
\item{glm_nml}{a nml (a list) for GLM config}
}
\description{
This function prints the specified values from the nml list for GLM.
}
\examples{
file = '../resources/glm.nml'
glm_nml <- read_nml(file)
pretty_nml(glm_nml)
}
\author{
Jordan S. Read
}

