% Generated by roxygen2 (4.0.2): do not edit by hand
\name{print.nml}
\alias{print.nml}
\title{prints R nml list with pretty formatting}
\usage{
\method{print}{nml}(glm_nml)
}
\arguments{
\item{glm_nml}{a nml (a list) for GLM config}
}
\description{
This function prints the specified values from the nml list for GLM.
}
\examples{
glm_nml <- read_nml()
print(glm_nml)
}
\author{
Jordan S. Read
}

