% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_field_obs.R
\name{read_field_obs}
\alias{read_field_obs}
\title{Read in field data into a data.frame}
\usage{
read_field_obs(file, var_name = "temp")
}
\arguments{
\item{file}{a file path for a supported field data format (currently, tsv,
csv, or RDS containing a data frame with Date, depth, wtemp or similar columns)}

\item{var_name}{Name of variable to look for in field_obs file. 
Should match a GLM simulation variable (see output from \code{\link{sim_vars}}).
Defaults to 'temp'}
}
\value{
a data.frame with DateTime, Depth, and wTemp columns
}
\description{
Read in field data into a data.frame
}
\examples{
# -- read in a .csv file --
file <- system.file('extdata', 'LakeMendota_field_data_hours.csv', package = 'glmtools')
field_data <- read_field_obs(file)

}
\seealso{
\link{get_temp}, \link{resample_sim}
}
\author{
Jordan S. Read
}
\keyword{methods}
