% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/read_field_stage.R
\name{read_field_stage}
\alias{read_field_stage}
\title{Read in a field stage file}
\usage{
read_field_stage(file)
}
\arguments{
\item{file}{Path to stage TSV or CSV file}
}
\description{
Reads a field file for stage. Field file must be either
a TSV (tab separated) or CSV (comma separated) file with
columns named "datetime" and "stage" (in meters). Datetime
format must be yyyy-mm-dd
}
\details{

}
\examples{
field_file = system.file('extdata/field_stage.csv',
                         package='glmtools')

data <- read_field_stage(field_file)

plot(data)
}
\author{
Luke Winslow
}

