% Generated by roxygen2 (4.0.1): do not edit by hand
\name{resample_to_field}
\alias{resample_to_field}
\title{match GLM water temperatures with field observations}
\usage{
resample_to_field(nc_file, field_file)
}
\arguments{
\item{nc_file}{a string with the path to the netcdf output from GLM}

\item{field_file}{a string with the path to the field observation file}
}
\value{
validation a data.frame with DateTime and temperature at depth
}
\description{
match GLM water temperatures with field observations
}
\examples{
nc_file <- system.file('extdata', 'output.nc', package = 'glmtools')
field_file <- system.file('extdata', 'field_data.tsv', package = 'glmtools')

temps <- resample_to_field(nc_file, field_file)
}
\author{
Jordan S. Read
}
\seealso{
\link{resample_time}, \link{get_temp}
}
\keyword{methods}

