% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/run_example_sim.R
\name{run_example_sim}
\alias{run_example_sim}
\title{Run example simulation}
\usage{
run_example_sim(sim_folder, verbose = TRUE)
}
\arguments{
\item{sim_folder}{the directory where simulation files will be copied/moved to.
Will use a temporary directory if missing}

\item{verbose}{should operations and output of GLM be shown}
}
\description{
Run example simulation
}
\examples{
run_example_sim()
\dontrun{
run_example_sim('~/Desktop/GLM_sims')
}
}
\author{
Jordan S. Read, Luke A. Winslow
}
\seealso{
\code{\link[GLMr]{run_glm}}
}
\keyword{methods}

