% Generated by roxygen2 (4.0.1): do not edit by hand
\name{run_glmOSx}
\alias{run_glmOSx}
\title{run GLM on a mac}
\usage{
run_glmOSx(sim_folder, verbose = TRUE)
}
\arguments{
\item{sim_folder}{the directory where simulation files are contained}

\item{field_file}{a string with the path to the field observation file}

\item{verbose}{should output of GLM be shown}
}
\description{
run GLM on a mac
}
\examples{
\dontrun{
sim_folder <- system.file('extdata/sim/', package = 'glmtools')
run_glmOSx(sim_folder)
out_file <- file.path(sim_folder,'output.nc')
nml_file <- system.file('extdata/sim', 'glm.nml', package = 'glmtools')
field_file <- system.file('extdata', 'field_data.tsv', package = 'glmtools')
plot_temp(file = out_file, fig_path = 'test_temp.png')
model_diagnostics(out_file, field_file, nml_file = nml_file,
   metrics = c('center.buoyancy','thermo.depth', 'schmidt.stability'),
   fig_path = 'test_diagnostic.png')
}
}
\author{
Jordan S. Read
}
\seealso{
\link{compare_to_field}, \link{resample_to_field}, \link{read_nml}, \link{get_metrics}
}
\keyword{methods}

