\name{runRemoteLA}
\alias{runRemoteLA}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Runs zipped files on remote Lake Analyzer server}
\description{This function runs zipped files on remote Lake Analyzer server}
\usage{
runRemoteLA(zipPath,destFldr)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{zipPath}{location of zipped files for Lake Analyzer run}
  \item{destFldr}{destination folder for zipped results files}
}
\value{
}
\author{
Jordan S. Read
}

\keyword{rGLM}
