\name{set.lke}
\alias{set.lke}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{sets value in .lke list}
\description{This function sets value in .lke list}
\usage{
set.lke(lke,argName,argVal)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{lke}{lke list}
  \item{argName}{lke list argument string}
  \item{argValue}{lke list argument value}
}
\value{
a modified lke list
}
\author{
Jordan S. Read
}

\keyword{rGLM}
