\name{set.nml}
\alias{set.nml}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{sets value in R nml list}
\description{This function sets the specified value in the nml list for GLM. All data types are converted tested against initial data type formats of logical, numerical, or character arrays}
\usage{
set.nml(nml,argName,argVal,argList)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{nml}{ an nml list}
  \item{argName}{ a valid nml parameter}
	\item{argVal}{ a valid nml value}
	\item{argList}{ a valid list of nml parameter names and their values (optional).}
}
\value{
list of nml parameters
}
\author{
Jordan S. Read
}

\keyword{rGLM}
