% Generated by roxygen2 (4.0.1): do not edit by hand
\name{set_lke}
\alias{set_lke}
\title{sets values in lke object}
\usage{
set_lke(lke, arg_name, arg_val)
}
\arguments{
\item{lke}{a nml (a list) for Lake Analyzer config}

\item{arg_name}{a string representing a valid field in the lke object}

\item{arg_val}{value for the valid field in lke specified by \code{arg_name}}
}
\value{
lke a modified lke object
}
\description{
This function sets values in lke object for Lake Analyzer config.
}
\examples{
lke <- init_lke()
lke <- set_lke(lke, arg_name='totalDep', arg_val=19.4)
}
\author{
Jordan S. Read
}
\seealso{
\link{init_lke}, \link{write_lke}
}

