% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_metrics.R
\name{sim_metrics}
\alias{sim_metrics}
\title{get possible metrics for comparing GLM outputs to field}
\usage{
sim_metrics(with_nml = FALSE)
}
\arguments{
\item{with_nml}{include metrics that require information in the nml file}
}
\value{
a character array of metrics that can be used
}
\description{
returns a character array of metric names that can be supported by \code{compare_to_field}. 
This function is only designed to handle calls to physical metrics that return a single value.
An example of this behavior is thermocline depth (but not water density for all depth-resolved measurements).
}
\examples{
sim_metrics()
sim_metrics(with_nml = TRUE)
}
\author{
Jordan S. Read
}
\seealso{
\link{compare_to_field}, \link{resample_to_field}, \link{read_nml}
}
\keyword{methods}

