% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sim_var_units.R
\name{sim_var_units}
\alias{sim_var_units}
\title{get units of variable from a GLM simulation}
\usage{
sim_var_units(file, var_name)
}
\arguments{
\item{file}{a string with the path to the netcdf output from GLM}

\item{var_name}{name of a valid variable from a simulation (see \code{\link{sim_vars}})}
}
\description{
get units of variable from a GLM simulation
}
\examples{
sim_folder <- run_example_sim(verbose = FALSE)
nc_file <- file.path(sim_folder, 'output.nc')
vars <- sim_vars(file = nc_file)
sim_var_longname(nc_file, vars[1])
sim_var_longname(nc_file, 'u_mean')
}
\author{
Jordan S. Read
}
\seealso{
\code{\link{sim_vars}}, \code{\link{sim_var_longname}}
}
\keyword{methods}

