% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sim_vars}
\alias{sim_vars}
\title{get list of variables from a GLM simulation}
\usage{
sim_vars(file)
}
\arguments{
\item{file}{a string with the path to the netcdf output from GLM}
}
\description{
get list of variables from a GLM simulation
}
\examples{
sim_folder <- run_example_sim(verbose = FALSE)
nc_file <- file.path(sim_folder, 'output.nc')
sim_vars(file = nc_file)
}
\author{
Stuart E. Jones, Jordan S. Read, Luke A. Winslow
}
\keyword{methods}

