\name{subsampleGLM}
\alias{subsampleGLM}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{gets wtr temperature observations at depths from GLM data.frame}
\description{This function gets wtr temperature observations at depths from GLM data.frame}
\usage{
subsampleGLM(GLM, sampleTime, sampleDepths)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{GLM}{GLM data.frame}
  \item{sampleTime}{times for subsampling}
  \item{sampleDepths}{depths for subsampling}
}
\value{
Water values for GLM output
}
\author{
Jordan S. Read & Luke A. Winslow
}

\keyword{rGLM}
