% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_metrics.R
\name{water.temperature}
\alias{water.temperature}
\title{mimic rLakeAnalyzer function}
\usage{
water.temperature(wtr)
}
\arguments{
\item{wtr}{a water temperature data.frame}
}
\details{
used to mimic an rLakeAnalyzer function, so that it will 
be recognized by a call to \code{\link{sim_metrics}}
}
\seealso{
\code{\link{sim_metrics}}, \code{\link{compare_to_field}}, \code{\link{validate_sim}}
}

