\name{write.bth}
\alias{write.bth}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{writes bth data.frame to .bth file}
\description{This function writes bth data.frame to .bth file}
\usage{
write.bth(bth,lakeName,folder)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{bth}{bth data.frame to be written to .bth file}
  \item{lakeName}{name of .bth file to be written}
  \item{folder}{folder where the .bth file will be written to}
}
\value{
}
\author{
Jordan S. Read
}

\keyword{rGLM}
