\name{write.lke}
\alias{write.lke}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{writes lke list to .lke file}
\description{This function writes lke list to .lke file}
\usage{
write.lke(lke,lakeName,folder)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{lke}{lke list to be written to .lke file}
  \item{lakeName}{name of .lke file to be written}
  \item{folder}{folder where the .lke file will be written to}
}
\value{
}
\author{
Jordan S. Read
}

\keyword{rGLM}
