\name{write.lvl}
\alias{write.lvl}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{writes lvl data.frame to .lvl file}
\description{This function writes lvl data.frame to .lvl file}
\usage{
write.lvl(lvl,lakeName,folder)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{lvl}{lvl data.frame to be written to .lvl file}
  \item{lakeName}{name of .lvl file to be written}
  \item{folder}{folder where the .lvl file will be written to}
}
\value{
}
\author{
Jordan S. Read
}

\keyword{rGLM}
