\name{write.nml}
\alias{write.nml}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{writes R nml list to an .nml text file}
\description{This function writes the specified values from the nml list for GLM to an .nml file.}
\usage{
	write.nml(nml,fileName,folder)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{nml}{ an nml list}
  \item{fileName}{file name for nml list to be written to}
  \item{folder}{ folder where text file will be written to}
}

\author{
Jordan S. Read
}

\keyword{rGLM}
