\name{write.wnd}
\alias{write.wnd}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{writes wnd data.frame to .wnd file}
\description{This function writes wnd data.frame to .wnd file}
\usage{
write.wnd(wnd,lakeName,folder)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{wnd}{wnd data.frame to be written to .wnd file}
  \item{lakeName}{name of .wnd file to be written}
  \item{folder}{folder where the .wnd file will be written to}
}
\value{
}
\author{
Jordan S. Read
}

\keyword{rGLM}
