\name{write.wtr}
\alias{write.wtr}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{writes wtr data.frame to .wtr file}
\description{This function writes wtr data.frame to .wtr file}
\usage{
write.wtr(wtr,lakeName,folder)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{wtr}{wtr data.frame to be written to .wtr file}
  \item{lakeName}{name of .wtr file to be written}
  \item{folder}{folder where the .wtr file will be written to}
}
\value{
}
\author{
Jordan S. Read
}

\keyword{rGLM}
