% Generated by roxygen2 (4.0.1): do not edit by hand
\name{write_bth}
\alias{write_bth}
\title{write Lake Analyzer *.bth file from GLM simulation}
\usage{
write_bth(glm_nml, lake_name = "lake", folder_out)
}
\arguments{
\item{glm_nml}{a nml (a list) for GLM config}

\item{lake_name}{a string for the name of the lake (used in *.lvl file naming)}

\item{folder_out}{a boolean for including ice thickness in surface height}
}
\description{
Creates a *.bth file for Lake Analyzer use from GLM simulation.  \cr
}
\examples{
nml_file <- system.file('extdata', 'glm.nml', package = 'glmtools')
folder_out <- system.file('extdata', package = 'glmtools')
glm_nml <- read_nml(file = nml_file)
write_bth(glm_nml, lake_name='lake', folder_out = folder_out)
}
\author{
Jordan S. Read
}
\seealso{
\link{write_wtr}, \link{write_lvl}, \link{get_wind}
}
\keyword{methods}

