% Generated by roxygen2 (4.0.1): do not edit by hand
\name{write_lke}
\alias{write_lke}
\title{write Lake Analyzer *.lke file from GLM simulation}
\usage{
write_lke(lke, lake_name = "lake", folder_out)
}
\arguments{
\item{lke}{a nml (a list) for GLM config}

\item{lake_name}{a string for the name of the lake (used in *.lke file naming)}

\item{folder_out}{a boolean for including ice thickness in surface height}
}
\description{
Creates a *.lke file for Lake Analyzer use from GLM simulation.  \cr
}
\examples{
lke <- init_lke()
lke <- set_lke(lke, arg_name = 'totalDep', arg_val = 19.4)
folder_out <- system.file('extdata', package = 'rGLM')
write_lke(lke, lake_name = 'lake', folder_out = folder_out)
}
\author{
Jordan S. Read
}
\seealso{
\link{write_wtr}, \link{write_lvl}, \link{init_lke}, \link{set_lke}
}
\keyword{methods}

