% Generated by roxygen2 (4.0.2): do not edit by hand
\name{write_lvl}
\alias{write_lvl}
\title{write Lake Analyzer *.lvl file from GLM simulation}
\usage{
write_lvl(glm_surface, glm_nml, lake_name = "lake", folder_out)
}
\arguments{
\item{glm_surface}{a data.frame with DateTime and surface elevation from from GLM}

\item{glm_nml}{a nml (a list) for GLM config}

\item{lake_name}{a string for the name of the lake (used in *.lvl file naming)}

\item{folder_out}{a boolean for including ice thickness in surface height}
}
\description{
Creates a *.lvl file from GLM simulation.  \cr
}
\examples{
file <- system.file('extdata', 'output.nc', package = 'glmtools')
nml_file <- system.file('extdata', 'glm.nml', package = 'glmtools')
glm_surface <- get_surface_height(file, ice.rm = TRUE)
folder_out <- system.file('extdata', package = 'glmtools')
glm_nml <- read_nml(nml_file)
write_lvl(glm_surface, glm_nml, lake_name = 'lake', folder_out = folder_out)
}
\author{
Jordan S. Read
}
\seealso{
\link{get_nml_value}, \link{read_nml}, \link{get_surface_height}
}
\keyword{methods}

