% Generated by roxygen2 (4.0.1): do not edit by hand
\name{write_wnd}
\alias{write_wnd}
\title{write Lake Analyzer *.wnd file from GLM simulation}
\usage{
write_wnd(glm_wind, lake_name = "lake", folder_out = "../Supporting Files/")
}
\arguments{
\item{glm_wnd}{a data.frame with DateTime and temperatures from GLM}

\item{lake_name}{a string for the name of the lake (used in *.lvl file naming)}

\item{folder_out}{a string for the output folder for the *.lvl file}
}
\description{
Creates a *.wnd file for Lake Analyzer use from GLM simulation.  \cr
}
\examples{
file = '../test/output.nc'
glm_wind <- get_wind(file)
write_wnd(glm_wind,lake_name='lake',folder_out='../resources/')
}
\author{
Jordan S. Read
}
\seealso{
\link{write_wtr}, \link{write_lvl}, \link{get_wind}
}
\keyword{methods}

