% Generated by roxygen2 (4.0.1): do not edit by hand
\name{write_wtr}
\alias{write_wtr}
\title{write Lake Analyzer *.wtr file from GLM simulation}
\usage{
write_wtr(glm_temp, lake_name = "lake", folder_out = "../resources/")
}
\arguments{
\item{glm_temp}{a data.frame with DateTime and temperatures from GLM}

\item{lake_name}{a string for the name of the lake (used in *.lvl file naming)}

\item{folder_out}{a boolean for including ice thickness in surface height}
}
\description{
Creates a *.wtr file for Lake Analyzer use from GLM simulation.  \cr
}
\examples{
file = '../test/output.nc'
z_out <- c(0,1,2,3,4,5,6)
glm_temp <- get_temp(file,reference='surface',z_out=z_out)
write_wtr(glm_temp,lake_name='lake',folder_out='../resources/')
}
\author{
Jordan S. Read
}
\seealso{
\link{write_lvl}, \link{get_temp}
}
\keyword{methods}

