\name{zipLA}
\alias{zipLA}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{zips up files for Lake Analyzer run}
\description{This function zips up files for Lake Analyzer run}
\usage{
zipLA(lakeName,folder,destFldr)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{lakeName}{name of files to be zipped}
  \item{folder}{folder where the files to be zipped reside in}
  \item{destFldr}{destination folder for zipped files}
}
\value{
}
\author{
Jordan S. Read
}

\keyword{rGLM}
