% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate_sim.R
\name{calibrate_sim}
\alias{calibrate_sim}
\title{Calibrates GLM-AED2 variables to improve fit between observed and simulated data}
\usage{
calibrate_sim(var = "temp", path, obs, nml.file = "glm3.nml",
  calib_setup = NULL, glmcmd = NULL, first.attempt = TRUE,
  period = NULL, scaling = "TRUE", method = "CMA-ES",
  metric = "RMSE", target.fit = 1.5, target.iter = 100,
  plotting = TRUE, output)
}
\arguments{
\item{var}{Character vector of valid variable names (see \code{\link{sim_vars}})}

\item{path}{String with the path to the GLM setup}

\item{obs}{CSV or TSV field data (see \link{resample_to_field}for format)}

\item{nml.file}{String of the glm-namelist file, default is 'glm3.nml'}

\item{calib_setup}{Data frame containing information regarding the calibration (see \link{get_calibration_setup})}

\item{glmcmd}{String containing the desired glm run command, default is GLM3r}

\item{first.attempt}{Boolean, if TRUE a nml-template will be created, set it to FALSE after your first calibration run; default is TRUE}

\item{period}{List that provides a start and a stop date for the simulation}

\item{scaling}{Boolean, if TRUE variable values will be scaled on the space (0,10), recommended for CMA-ES, default is TRUE}

\item{method}{String of the optimization method, default is 'CMA-ES' (Hansen 2009), alternatively you can also use 'Nelder-Mead'}

\item{metric}{String of the calibration fit metric, default is RMSE}

\item{target.fit}{Double of your preferred fit, calibration will stop after reaching that; default is 1.5}

\item{target.iter}{Double of maximum amount of iterations, default is 150}

\item{plotting}{Boolean, if TRUE plots all results as heat maps}

\item{output}{Character array of the output folder path}
}
\description{
Starts a calibration run with multiple iterations using a GLM-AED2 setup. At the end a output csv.-file of all iterations and a 
diagnostic plot are created.
}
\examples{
calib_setup <- get_calibration_setup()
print(calib_setup)

#Essential variables
var = 'temp' # variable to apply the calibration procedure
path = getwd() # simulation path
obs = read_field_obs('bcs/sparkling_lter_temp.csv') # observed field data
nml.file = 'glm3.nml' # name of nml-file, if different than the default version the system command needs to be changed as well
glmcmd = "glm" # command to be used, default applies GLM3r
#Optional variables
first.attempt = TRUE # if TRUE, deletes all local csv-files that stores the outcome of previous calibration runs
period = use get_calib_periods.R to create calibration/validation periods, i.e. period = get_calib_periods(nml = nml.file, ratio = 1)
scaling = TRUE # scaling should be TRUE for CMA-ES
method = 'CMA-ES' # Choose the optimization method, either Covariance Matrix Adaption - Evolution Strategy ('CMA-ES') or Nelder-Mead
metric = 'RMSE' # Root-mean square error
target.fit = 1.5 # refers to a target fit of 1.5 degrees Celsius (algorithm stops after reaching 1.5 degree Celsius, set target.fit = -Inf to run until it reaches target.iter only)
target.iter = 150 # refers to a maximum run of 150 calibration iterations (algorithm stops after doing 150 function evaluations)
plotting = TRUE
output = file.path(sim_folder, 'output/output.nc')

calibrate_sim(var, path, obs, nml.file, calib_setup, glmcmd, first.attempt, period, scaling, method, metric, target.fit, target.iter)

}
\seealso{
\code{\link{get_calib_setup}, \link{get_calib_periods}, \link{get_calib_init_validation}}
}
\author{
Robert Ladwig, Tadhg Moore
}
\keyword{methods}
