% Generated by roxygen2 (4.0.1): do not edit by hand
\name{compare_to_field}
\alias{compare_to_field}
\title{compare metric for GLM vs field observations}
\usage{
compare_to_field(nc_file, field_file, nml_file, metric, as_value = FALSE,
  na.rm = TRUE)
}
\arguments{
\item{nc_file}{a string with the path to the netcdf output from GLM}

\item{field_file}{a string with the path to the field observation file}

\item{nml_file}{a string with the path to the nml file (optional)}

\item{metric}{a string representing a physical metric. Should be a rLakeAnalyzer function or other function.}

\item{as_value}{a boolean for calculating RMSE (F) or returning all values (T)}

\item{na.rm}{a boolean for remove NAs for RMSE calculation (only used if as_values == F)}
}
\value{
a RMSE (in native units) for the comparison, or DateTime and all values as a data.frame (if as_values == T)
}
\description{
compare metric for GLM vs field observations, but must have more than 3 matching time points.
This function is only designed to handle calls to physical metrics that return a single value.
An example of this behavior is thermocline depth (but not water density for all depth-resolved measurements).
}
\examples{
nc_file <- system.file('extdata', 'output.nc', package = 'glmtools')
field_file <- system.file('extdata', 'field_data.tsv', package = 'glmtools')

CB_rmse <- compare_to_field(nc_file, field_file, metric = 'center.buoyancy')
CB_values <- compare_to_field(nc_file, field_file,
                          metric = 'center.buoyancy', as_value = TRUE)

thermo_values <- compare_to_field(nc_file, field_file,
                          metric = 'thermo.depth', as_value = TRUE)

\dontrun{
# -- an nml file is necessary when functions require hypsographic information
values <- compare_to_field(nc_file, field_file,
                          metric = 'schmidt.stability', as_value = TRUE)
# -- will fail
nml_file <- system.file('extdata', 'glm.nml', package = 'glmtools')
values <- compare_to_field(nc_file, field_file, nml_file,
                          metric = 'schmidt.stability', as_value = TRUE)
# -- will succeed

# -- metrics can only be calculated by functions that are available to this environment
values <- compare_to_field(nc_file, field_file, metric = 'calc.fols', as_value = TRUE)
# -- will fail
}
}
\author{
Jordan S. Read
}
\seealso{
\link{resample_time}, \link{resample_to_field}
}
\keyword{methods}

