% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_sim_var.R
\name{convert_sim_var}
\alias{convert_sim_var}
\title{convert an existing simulation variable into a different one}
\usage{
convert_sim_var(nc_file = "output.nc", convert, name, unit, longname = name)
}
\arguments{
\item{nc_file}{the netcdf output file}

\item{convert}{a function to convert variable based on other variables or offsets}

\item{name}{the name for the new variable}

\item{unit}{the units for the new variable}

\item{longname}{the longname for the new variable}
}
\description{
allows you to provide specific variable or unit conversions to create a new variable 
that can be used with \code{\link{plot_var_compare}}, and others.
}

