% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_sim_var.R
\name{convert_sim_var}
\alias{convert_sim_var}
\title{convert an existing simulation variable into a different one}
\usage{
convert_sim_var(nc_file = "output.nc", ..., unit = "", longname = "",
  overwrite = FALSE)
}
\arguments{
\item{nc_file}{the netcdf output file}

\item{\dots}{an expression to convert variable based on other variables or offsets}

\item{unit}{the units for the new variable}

\item{longname}{the longname for the new variable}

\item{overwrite}{logical overwrite existing variable data}
}
\description{
allows you to provide specific variable or unit conversions to create a new variable 
that can be used with \code{\link{plot_var_compare}}, and others.
}
\examples{
\dontrun{
sim_folder <- run_example_sim(verbose = FALSE)
nc_file <- file.path(sim_folder, 'output.nc')
convert_sim_var(nc_file, tempF = temp/5*9+32, unit='degF',longname='temperature degrees Farenheit')
plot_var(nc_file, 'tempF')
convert_sim_var(nc_file, crazy_var = temp-u_mean*1000)
plot_var(nc_file, 'crazy_var')

temp2f <- function(c) c/5*9+32
convert_sim_var(nc_file, tempf = temp2f(temp), unit='degF',longname='temperature degrees Farenheit')
}

}
