% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_calib_setup.R
\name{get_calib_setup}
\alias{get_calib_setup}
\title{Creates an example setup for a calibration run}
\usage{
get_calib_setup()
}
\description{
Creates dataframe containing the parameters for calibration (string) and their respective lower boundaries, upper boundaries 
and initial values
}
\examples{
calib_setup <- get_calibration_setup()
}
\author{
Robert Ladwig, Tadhg Moore
}
\keyword{methods}
