% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get_evaporation}
\alias{get_evaporation}
\title{get evaporation from GLM simulation}
\usage{
get_evaporation(file)
}
\arguments{
\item{file}{a string with the path to the netcdf output from GLM}
}
\value{
a data.frame with DateTime and evaporation (in mm/day)
}
\description{
Creates a data.frame with DateTime and E (mm/day).  \cr
}
\examples{
file <- system.file('extdata', 'output.nc', package = 'glmtools')
evap <- get_evaporation(file)
plot(evap)
}
\author{
Jordan S. Read
}
\keyword{methods}

