% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get_hypsography}
\alias{get_hypsography}
\title{retrieve hypsography information from glm_nml object or file}
\usage{
get_hypsography(glm_nml = NULL, file = NULL)
}
\arguments{
\item{glm_nml}{a nml (a list) for GLM config}

\item{file}{a string with the path to the GLM glm.nml file}
}
\value{
glm_bth a data.frame with \code{Depths} and \code{Areas}
}
\description{
Retrieves hypsography information from glm_nml object or file  \cr
}
\examples{
glm_nml <- read_nml('../resources/glm.nml')
write_bth(glm_nml, lake_name='lake', folder_out='../resources/')
}
\author{
Jordan S. Read
}
\seealso{
\link{write_lvl}, \link{read_nml}, \link{get_nml_value}
}
\keyword{methods}

