% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_hypsography.R
\name{get_hypsography}
\alias{get_hypsography}
\title{retrieve hypsography information}
\usage{
get_hypsography(glm_nml, file)
}
\arguments{
\item{glm_nml}{a nml (a list) for GLM config}

\item{file}{a string with the path to the GLM glm.nml file}
}
\value{
glm_bth a data.frame with \code{Depths} and \code{Areas}
}
\description{
Retrieves hypsography information from glm_nml object or file.  \cr
Hypsography is the relationship between depth and area of a lake.
'Depth' is referenced from the lake surface and downward values are positive in meters. Areas are in square meters.
}
\examples{
sim_folder <- run_example_sim(verbose = FALSE)
nml_file <- file.path(sim_folder, 'glm2.nml')
glm_nml <- read_nml(nml_file)
get_hypsography(glm_nml)
get_hypsography(file = nml_file)
}
\author{
Jordan S. Read
}
\seealso{
\link{read_nml}, \link{get_nml_value}
}
\keyword{methods}

