% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_ice.R
\name{get_ice}
\alias{get_ice}
\title{get ice depth from GLM simulation}
\usage{
get_ice(file = "output.nc", snow.rm = TRUE, ...)
}
\arguments{
\item{file}{a string with the path to the netcdf output from GLM}

\item{snow.rm}{a boolean for ignoring snow depth in the calculation of ice thickness}

\item{...}{additional arguments passed to \code{\link{resample_sim}}}
}
\value{
a data.frame with DateTime and ice (in meters)
}
\description{
Creates a data.frame with DateTime and ice.  \cr
This function sums the thickness of the clear ice, white ice,
and optionally, the thickness of snow for each timestep in the GLM model.
}
\examples{
sim_folder <- run_example_sim(verbose = FALSE)
nc_file <- file.path(sim_folder, 'output.nc')
ice <- get_ice(nc_file)
ice_and_snow <- get_ice(nc_file, snow.rm = FALSE)
plot(ice)
points(ice_and_snow, col = "red")
}
\author{
Luke A. Winslow, Jordan S. Read
}
\keyword{methods}

