% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw.R
\name{get_raw}
\alias{get_raw}
\title{get raw data from GLM simulation}
\usage{
get_raw(file = "output.nc", param_name)
}
\arguments{
\item{file}{a string with the path to the netcdf output from GLM}

\item{param_name}{Name of parameter requested (see details for list)}
}
\value{
the raw matrix or vector from the NC file
}
\description{
Returns the raw data in a GLM simulation nc file. \cr
}
\details{
NS\cr
hice\cr
hsnow\cr
hwice\cr
precip\cr
evap\cr
I_0\cr
wind\cr
Tot_V\cr
V\cr
salt\cr
temp\cr
rho\cr
rad\cr
ext_coef\cr
z\cr
}
\examples{
sim_folder <- run_example_sim(verbose = FALSE)
nc_file <- file.path(sim_folder, 'output/output.nc')
wind <- get_raw(nc_file, 'wind') #raw wind data
}
\seealso{
\code{\link{sim_vars}}
}
\author{
Luke A. Winslow, Jordan S. Read
}
\keyword{methods}
