% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_raw}
\alias{get_raw}
\title{get raw data from GLM simulation}
\usage{
get_raw(file, param_name)
}
\arguments{
\item{file}{a string with the path to the netcdf output from GLM}

\item{param_name}{Name of parameter requested}
}
\value{
the raw matrix or vector from the NC file
}
\description{
Returns the raw data in a GLM simulation nc file. \cr
}
\examples{
file <- system.file('extdata', 'output.nc', package = 'glmtools')
wind <- get_raw(file, 'temp') #raw elevation data
}
\author{
Luke A. Winslow, Jordan S. Read
}
\keyword{methods}

