% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get_temp}
\alias{get_temp}
\title{get water temperatures from a GLM simulation}
\usage{
get_temp(file, reference = "bottom", z_out)
}
\arguments{
\item{file}{a string with the path to the netcdf output from GLM}

\item{reference}{a string which specifies the vertical reference ('surface' or 'bottom')}

\item{z_out}{a vector of depths for temperature output (in meters)}
}
\value{
a data.frame with DateTime and temperature at depth
}
\description{
Creates a data.frame with DateTime and temperatures (in deg C).  \cr
Temperatures that are sampled out of the GLM output are taken relative
to the surface (\code{reference = 'surface'}) or the bottom of the lake
(\code{reference = 'bottom'}).
}
\examples{
file <- system.file('extdata', 'output.nc', package = 'glmtools')
temp_surf <- get_temp(file, reference = 'surface', z_out = c(0,1,2))
temp_bot <- get_temp(file, reference = 'bottom', z_out = c(0,1,2))
}
\author{
Jordan S. Read, Luke A. Winslow
}
\keyword{methods}

