% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get_temp}
\alias{get_temp}
\title{get ice depth from GLM simulation}
\usage{
get_temp(file, reference = "bottom", z_out)
}
\arguments{
\item{file}{a string with the path to the netcdf output from GLM}

\item{reference}{a string which specifies the vertical reference ('surface' or 'bottom')}

\item{z_out}{a vector of depths for temperature output (in meters)}
}
\value{
a data.frame with DateTime and temperature at depth
}
\description{
Creates a data.frame with DateTime and ice.  \cr
}
\examples{
file = '../test/output.nc'
temp_surf <- get_temp(file,reference='surface',z_out=c(0,1,2))
temp_bot <- get_temp(file,reference='bot',z_out=c(0,1,2))
}
\author{
Luke A. Winslow, Jordan S. Read
}
\keyword{methods}

