% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get_wind}
\alias{get_wind}
\title{get wind speed from GLM simulation}
\usage{
get_wind(file)
}
\arguments{
\item{file}{a string with the path to the netcdf output from GLM}
}
\value{
a data.frame with DateTime and wind
}
\description{
Creates a data.frame with DateTime and wind.  \cr
}
\examples{
file = '../test/output.nc'
ice <- get_wind(file)
}
\author{
Luke A. Winslow, Jordan S. Read
}
\keyword{methods}

