% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_wind.R
\name{get_wind}
\alias{get_wind}
\title{get wind speed from GLM simulation}
\usage{
get_wind(file, ...)
}
\arguments{
\item{file}{a string with the path to the netcdf output from GLM}

\item{...}{additional arguments passed to \code{\link{resample_sim}}}
}
\value{
a data.frame with DateTime and wind
}
\description{
Creates a data.frame with DateTime and wind.  \cr
}
\examples{
sim_folder <- run_example_sim(verbose = FALSE)
nc_file <- file.path(sim_folder, 'output.nc')
wind <- get_wind(file = nc_file)
}
\author{
Luke A. Winslow, Jordan S. Read
}
\keyword{methods}

