% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot_compare_stage.R
\name{plot_compare_stage}
\alias{plot_compare_stage}
\title{Plot simulated and observed stage data}
\usage{
plot_compare_stage(nc_file, field_file, ...)
}
\arguments{
\item{nc_file}{NetCDF model output file}

\item{field_file}{CSV or TSV field data file (see \link{resample_to_field} for format)}

\item{...}{Additional paramters supplied to the \code{plot} function}
}
\description{
Plot simulated and observed stage data
}
\examples{
sim_folder <- run_example_sim(verbose = FALSE)
nc_file <- file.path(sim_folder, 'output.nc')

field_file <- file.path(sim_folder, 'field_stage.csv')

plot_compare_stage(nc_file, field_file) ##makes a plot!
}

