% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot_meteo.R
\name{plot_meteo}
\alias{plot_meteo}
\title{plot meterological drivers from a csv file}
\usage{
plot_meteo(nml_file, fig_path = FALSE)
}
\arguments{
\item{nml_file}{a string with the path to the GLM glm2.nml file, or
\code{'template'} for loading the GLM template nml file with GLMr (default)}

\item{fig_path}{FALSE if plot to screen, string path if save plot as .png}
}
\description{
plot meterological drivers from a csv file
}
\examples{
nml_file <- nml_template_path()

plot_meteo(nml_file)
}
\author{
Jordan S. Read, Luke A. Winslow
}
\seealso{
\code{\link{read_nml}}
}
\keyword{methods}

