% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot_meteo}
\alias{plot_meteo}
\title{plot meterological drivers from a csv file}
\usage{
plot_meteo(nml_file, fig_path = FALSE)
}
\arguments{
\item{nml_file}{a string with the path to the nml file for the simulation}

\item{fig_path}{FALSE if plot to screen, string path if save plot as .png}
}
\description{
plot meterological drivers from a csv file
}
\examples{
nml_file <- nml_template_path()

plot_meteo(nml_file)
}
\author{
Jordan S. Read, Luke A. Winslow
}
\seealso{
\link{read_nml}
}
\keyword{methods}

