% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_meteo.R
\name{plot_meteo}
\alias{plot_meteo}
\title{Plot meterological drivers from a csv file}
\usage{
plot_meteo(met_file, xmin = NA, xmax = NA, fig_path = FALSE, ...)
}
\arguments{
\item{met_file}{Options: 1) file path to glm2.nml file. Date limits pulled from .nml file
2) file path to .csv file, with Date values in the first column
3) data.frame object, with Date values in the first column}

\item{xmin}{Optional, start date for xaxis}

\item{xmax}{Optional, end date for xaxis}

\item{fig_path}{Logical; F if plot to screen, string path if save plot as .png}

\item{...}{additional arguments passed to \code{\link[ggplot2:ggsave]{ggplot2:ggsave}}}
}
\description{
Plot meterological drivers from a csv file
}
\examples{
met_file <- nml_template_path()

plot_meteo(met_file)
}
\seealso{
\code{\link{read_nml}}
}
\author{
Jordan S. Read, Luke A. Winslow
}
\keyword{methods}
