% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot_temp.R
\name{plot_temp}
\alias{plot_temp}
\title{plot water temperatures from a GLM simulation}
\usage{
plot_temp(file, fig_path = F, ...)
}
\arguments{
\item{file}{a string with the path to the netcdf output from GLM}

\item{fig_path}{F if plot to screen, string path if save plot as .png}

\item{reference}{a string for 'surface' or 'bottom'}

\item{num_cells}{number of vertical cells to use for heatmap}

\item{add}{F if create new figure, T if add to existing}
}
\description{
plot water temperatures from a GLM simulation
}
\examples{
sim_folder <- run_example_sim(verbose = FALSE)
nc_file <- file.path(sim_folder, 'output.nc')
plot_temp(file = nc_file, fig_path = FALSE)
\dontrun{
plot_temp(file = nc_file, fig_path = '../test_figure.png')
}
}
\author{
Jordan S. Read, Luke A. Winslow
}
\seealso{
\code{\link{get_temp}}
}
\keyword{methods}

