% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot_temp}
\alias{plot_temp}
\title{plot water temperatures from a GLM simulation}
\usage{
plot_temp(file, reference = "surface", num_cells = 100, fig_path = F)
}
\arguments{
\item{file}{a string with the path to the netcdf output from GLM}

\item{refence}{a string for 'surface' or 'bottom'}

\item{num_cells}{number of vertical cells to use for heatmap}

\item{fig_path}{F if plot to screen, string path if save plot as .png}
}
\description{
plot water temperatures from a GLM simulation
}
\examples{
file <- system.file('extdata', 'output.nc', package = 'glmtools')
plot_temp(file = file, fig_path = FALSE)
plot_temp(file = file, fig_path = '../test_figure.png')
}
\author{
Jordan S. Read, Luke A. Winslow
}
\seealso{
\link{get_temp}
}
\keyword{methods}

