% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot_compare_heatmaps.R
\name{plot_temp_compare}
\alias{plot_temp_compare}
\title{Plot matching heatmaps for modeled and observed temp}
\usage{
plot_temp_compare(nc_file, field_file, ...)
}
\arguments{
\item{nc_file}{Netcdf model output file}

\item{field_file}{CSV or TSV field data file (see \link{resample_to_field} for format)}
}
\description{
Plot matching heatmaps for modeled and observed temp
}
\examples{
sim_folder <- run_example_sim(verbose = FALSE)
nc_file <- file.path(sim_folder, 'output.nc')
nml_file <- file.path(sim_folder, 'glm2.nml')
field_file <- file.path(sim_folder, 'field_data.tsv')

plot_temp_compare(nc_file, field_file) ##makes a plot!
}
\author{
Luke Winslow
}
\seealso{
Internally uses \link{get_temp} and \link{resample_to_field}
}

