% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_validate_profiles.R
\name{plot_validate_profiles}
\alias{plot_validate_profiles}
\title{Plot validation and model temperature profiles}
\usage{
plot_validate_profiles(nc_file, field_file, fig_path = FALSE, ...)
}
\arguments{
\item{nc_file}{a string with the path to the netcdf output from GLM}

\item{field_file}{a string with the path to the field observation file}

\item{fig_path}{FALSE if plot to screen, string path if save plot as .png. 
If argument is not used, plotting is skipped}

\item{...}{optional arguments passed to \code{resample_to_field}}
}
\examples{
sim_folder <- run_example_sim(verbose = FALSE)
nc_file <- file.path(sim_folder, 'output.nc')
nml_file <- file.path(sim_folder, 'glm2.nml')
field_file <- file.path(sim_folder, 'field_data.tsv')

#  create a multiple metric diagnostic fig within R:
plot_validate_profiles(nc_file, field_file, fig_path = FALSE, method = 'interp')                        
}
\seealso{
\link{validate_sim}, \link{resample_to_field}
}
\author{
Luke A. Winslow, Jordan S. Read
}
\keyword{methods}
