% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot_var.R
\name{plot_var}
\alias{plot_var}
\title{plot variable from a GLM simulation}
\usage{
plot_var(file, var_name, col_lim, reference = "surface", num_cells = 100,
  fig_path = F, add = F, bar_title = NULL)
}
\arguments{
\item{file}{a string with the path to the netcdf output from GLM}

\item{var_name}{a string for a valid variable name (see \code{\link{sim_vars}})}

\item{col_lim}{a numeric array of length 2 that specifies the min and max value for color bar}

\item{reference}{a string for 'surface' or 'bottom'}

\item{num_cells}{number of vertical cells to use for heatmap}

\item{fig_path}{F if plot to screen, string path if save plot as .png}

\item{add}{F if create new figure, T if add to existing}

\item{bar_title}{NULL if use \code{\link{sim_var_longname}}, or specify as a string to name plot variable}
}
\description{
plot variable from a GLM simulation
}
\examples{
sim_folder <- run_example_sim(verbose = FALSE)
nc_file <- file.path(sim_folder, 'output.nc')
vars <- sim_vars(file = nc_file)
plot_var(nc_file, 'u_mean')
\dontrun{
# need to specify a valid .nc file here:
plot_var(file = fabm_sim_nc.nc,
var_name = 'aed_oxygen_oxy',
col_lim=c(0,400),
bar_title = 'Dissolved Oxygen',
fig_path = 'aed_out.png')
}
}
\author{
Jordan S. Read, Luke A. Winslow
}
\seealso{
\code{\link{get_temp}}, \code{\link{sim_var_longname}}, \code{\link{sim_vars}}
}
\keyword{methods}

