% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot_var}
\alias{plot_var}
\title{plot water temperatures from a GLM simulation}
\usage{
plot_var(file, var_name, col_lim, reference = "surface", num_cells = 100,
  fig_path = F, add = F)
}
\arguments{
\item{file}{a string with the path to the netcdf output from GLM}

\item{var_name}{a string for a valid variable name}

\item{col_lim}{a numeric array of length 2 that specifies the min and max value for color bar}

\item{reference}{a string for 'surface' or 'bottom'}

\item{num_cells}{number of vertical cells to use for heatmap}

\item{fig_path}{F if plot to screen, string path if save plot as .png}

\item{add}{F if create new figure, T if add to existing}
}
\description{
plot water temperatures from a GLM simulation
}
\author{
Jordan S. Read, Luke A. Winslow
}
\seealso{
\link{get_temp}
}
\keyword{methods}

