% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot_var.R
\name{plot_var}
\alias{plot_var}
\title{plot variables from a GLM simulation}
\usage{
plot_var(file, var_name, fig_path = F, reference = "surface", ...)
}
\arguments{
\item{file}{a string with the path to the netcdf output from GLM}

\item{var_name}{a character vector of valid variable names (see \code{\link{sim_vars}})}

\item{fig_path}{F if plot to screen, string path if save plot as .png}

\item{reference}{'surface' or 'bottom'. Only used for heatmap plots.}

\item{...}{additional arguments passed to \code{par()}}
}
\description{
plot variables from a GLM simulation
}
\examples{
sim_folder <- run_example_sim(verbose = FALSE)
nc_file <- file.path(sim_folder, 'output.nc')
vars <- sim_vars(file = nc_file)
plot_var(nc_file, 'u_mean')
\dontrun{
# need to specify a valid .nc file here:
plot_var(file = fabm_sim_nc.nc,
var_name = 'aed_oxygen_oxy',
col_lim=c(0,400),
bar_title = 'Dissolved Oxygen',
fig_path = 'aed_out.png')
}
}
\author{
Jordan S. Read, Luke A. Winslow
}
\seealso{
\code{\link{get_temp}}, \code{\link{sim_var_longname}},
\code{\link{sim_vars}}, \code{\link{plot_temp}},  \code{\link{get_var}},
}
\keyword{methods}

