% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pretty_nml}
\alias{pretty_nml}
\title{prints R nml list with pretty formatting}
\usage{
pretty_nml(glm_nml)
}
\arguments{
\item{glm_nml}{a nml (a list) for GLM config}
}
\description{
This function prints the specified values from the nml list for GLM.
}
\examples{
nml_file <- system.file('extdata', 'glm.nml', package = 'glmtools')
glm_nml <- read_nml(nml_file)
pretty_nml(glm_nml)
}
\author{
Jordan S. Read
}

