% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_field_obs.R
\name{read_field_obs}
\alias{read_field_obs}
\title{read in field data into a data.frame}
\usage{
read_field_obs(file)
}
\arguments{
\item{file}{a file path for a supported field data format (currently, tsv or csv with Date, depth, wtemp or similar columns)}
}
\value{
a data.frame with DateTime, Depth, and wTemp columns
}
\examples{
# -- read in a .csv file --
file <- system.file('extdata', 'field_data.csv', package = 'glmtools')
field_data <- read_field_obs(file)

# -- read in a .tsv file --
file <- system.file('extdata', 'field_data.tsv', package = 'glmtools')
field_data <- read_field_obs(file)
}
\author{
Jordan S. Read
}
\seealso{
\link{get_temp}, \link{resample_sim}
}
\keyword{methods}

