% Generated by roxygen2 (4.0.1): do not edit by hand
\name{read_nml}
\alias{read_nml}
\title{read in a GLM simulation *.nml file}
\usage{
read_nml(file)
}
\arguments{
\item{file}{a string with the path to the GLM glm.nml file}
}
\value{
glm_nml a nml (a list) for GLM config
}
\description{
read in a GLM simulation *.nml file and create a list.  \cr
}
\examples{
nml_file <- system.file('extdata', 'glm.nml', package = 'rGLM')
glm_nml <- read_nml(file)
pretty_nml(glm_nml)
}
\author{
Jordan S. Read
}
\seealso{
\link{get_nml_value}
}
\keyword{methods}

