% Generated by roxygen2 (4.0.2): do not edit by hand
\name{resample_sim}
\alias{resample_sim}
\title{get subset of time from a generic timeseries data.frame}
\usage{
resample_sim(df, t_out, method = "match", precision = "day")
}
\arguments{
\item{df}{a data.frame with DateTime and potentially other columns}

\item{t_out}{a vector of POSIXct dates for matching to df$DateTime}

\item{method}{'match' for exact match or 'interp' for temporal interpolated (not yet supported)}

\item{precision}{matching precision (must be 'secs', 'mins','hours', or 'days')}
}
\value{
a data.frame with DateTime other original columns, resampled according to t_out
}
\description{
resamples the input data.frame to only have rows corresponding to matches between
df$DateTime and t_out. Both df$DateTime and t_out are of type POSIXct, and the
precision of the match is passed in through the \code{precision} argument.
\emph{The order of t_out}, not df$DateTime is retained.
}
\examples{
file <- system.file('extdata', 'output.nc', package = 'glmtools')
temp_surf <- get_temp(file, reference = 'surface', z_out = c(0,1,2))
t_out <- as.POSIXct(c("2011-04-01", "2011-06-14", "2011-04-05", "2011-07-28"))
temp_out <- resample_sim(df = temp_surf, t_out = t_out, precision = 'day')
}
\author{
Jordan S. Read
}
\seealso{
\link{get_temp}, \link{get_wind}, \link{get_surface_height}, \link{get_evaporation}, \link{get_ice}
}
\keyword{methods}

