% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/resample_to_field.R
\name{resample_to_field}
\alias{resample_to_field}
\title{match GLM water temperatures with field observations}
\usage{
resample_to_field(nc_file, field_file, method = "match", precision = "days")
}
\arguments{
\item{nc_file}{a string with the path to the netcdf output from GLM}

\item{field_file}{a string with the path to the field observation file}

\item{method}{'match' for exact match or 'interp' for temporal interpolation}

\item{precision}{matching precision (must be 'secs', 'mins','hours', or 'days')}
}
\value{
validation a data.frame with DateTime and temperature at depth
}
\description{
match GLM water temperatures with field observations
}
\examples{
sim_folder <- run_example_sim(verbose = FALSE)
nc_file <- file.path(sim_folder, 'output.nc')
field_file <- file.path(sim_folder, 'field_data.tsv')

temps <- resample_to_field(nc_file, field_file)
buoy_file <- system.file('extdata', 'buoy_data.csv', package = 'glmtools')
temps <- resample_to_field(nc_file, buoy_file, precision = 'hours')
}
\author{
Jordan S. Read
}
\seealso{
\link{resample_sim}, \link{get_temp}
}
\keyword{methods}

