% Generated by roxygen2 (4.0.1): do not edit by hand
\name{run_analyzer}
\alias{run_analyzer}
\title{Runs Lake Analyzer on remote server for GLM outputs}
\usage{
run_analyzer(lake_name = "lake", folder = "../resources/",
  folder_out = "../Supporting Files")
}
\arguments{
\item{lake_name}{a string for the name of the lake (used in *.lke file naming)}

\item{folder}{a string for the folder path of the Lake Analyzer files (e.g., *.lke)}

\item{folder_out}{a string for the destination of results output}
}
\value{
path_out a string with the file path to the output results
}
\description{
This function runs Lake Analyzer server according to the files created (e.g., *.lke, *.wtr, etc).  \cr
}
\examples{
file_path <- run_analyzer(lake_name='lake',folder='../resources/',
                         folder_out='../resources/')
}
\author{
Luke A. Winslow, Jordan S. Read
}
\seealso{
\link{write_lke}, \link{init_lke}, \link{set_lke}
}
\keyword{methods}

