% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sim_vars}
\alias{sim_vars}
\title{get list of variables from a GLM simulation}
\usage{
sim_vars(file)
}
\arguments{
\item{file}{a string with the path to the netcdf output from GLM}
}
\description{
get list of variables from a GLM simulation
}
\examples{
nc_file<-system.file('extdata','output.nc',package='glmtools')
sim_vars(file = nc_file)
}
\author{
Jordan S. Read, Luke A. Winslow, Stuart E. Jones
}
\keyword{methods}

