% Generated by roxygen2 (4.0.2): do not edit by hand
\name{summarize_sim}
\alias{summarize_sim}
\title{creates GLM simulation summary outputs}
\usage{
summarize_sim(file)
}
\arguments{
\item{file}{a string with the path to the netcdf output from GLM}
}
\value{
a data.frame with DateTime and temperature at depth
}
\description{
creates GLM simulation summary outputs
}
\author{
Jordan S. Read, Luke A. Winslow
}
\keyword{methods}

