% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_sim.R
\name{summarize_sim}
\alias{summarize_sim}
\title{Creates GLM simulation summary outputs}
\usage{
summarize_sim(file, sim_outputs, fig_path)
}
\arguments{
\item{file}{a string with the path to the netcdf output from GLM}

\item{sim_outputs}{a character vector of outputs.}

\item{fig_path}{FALSE if plot to screen, string path if save plot as .png
See \code{get_}, timeseries (as in temp, evaporation, etc)}
}
\description{
Creates GLM simulation summary outputs
}
\examples{
\dontrun{
sim_folder <- run_example_sim(verbose = FALSE)
nc_file <- file.path(sim_folder, 'output/output.nc')
summarize_sim(nc_file)

summarize_sim(nc_file, sim_outputs = c('temp','surface_height','wind'))
}
}
\seealso{
\link{get_temp}, \link{get_surface_height}, \link{get_evaporation}
}
\author{
Jordan S. Read, Luke A. Winslow
}
\keyword{methods}
