% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/validate_sim.R
\name{validate_sim}
\alias{validate_sim}
\title{run diagnostics on model results vs observations}
\usage{
validate_sim(nc_file, field_file, metrics, fig_path, report = F, ...)
}
\arguments{
\item{nc_file}{a string with the path to the netcdf output from GLM}

\item{field_file}{a string with the path to the field observation file}

\item{metrics}{a string or vector of strings representing physical metrics.
Should be a rLakeAnalyzer function or other valid function.}

\item{fig_path}{FALSE if plot to screen, string path if save plot as .png.
If argument is not used, plotting is skipped}

\item{report}{default as FALSE, so no stats are returned. If TRUE, return
summary stats for each metric used.}

\item{...}{additional arguments passed to compare_to_field (such as na.rm or nml_file)}
}
\value{
a report or figures comparing modeled results to observations
}
\description{
run diagnostics on model results vs observations
}
\examples{
sim_folder <- run_example_sim(verbose = FALSE)
nc_file <- file.path(sim_folder, 'output.nc')
nml_file <- file.path(sim_folder, 'glm2.nml')
field_file <- file.path(sim_folder, 'field_data.tsv')

#  create a multiple metric diagnostic fig within R:
validate_sim(nc_file, field_file, nml_file = nml_file,
                          metrics = c('thermo.depth', 'schmidt.stability'),
                          fig_path = FALSE)

# write the fig out to file:
validate_sim(nc_file, field_file, nml_file = nml_file, fig_path = '../test_fig.png',
                          metrics = c('thermo.depth', 'schmidt.stability'))

# return the diagnostic info but avoid the plot:
validate_sim(nc_file, field_file, nml_file = nml_file, report = TRUE,
                          metrics = c('thermo.depth', 'schmidt.stability'))
}
\author{
Jordan S. Read
}
\seealso{
\link{compare_to_field}, \link{resample_to_field}, \link{read_nml}, \link{sim_metrics}
}
\keyword{methods}

