% Generated by roxygen2 (4.0.1): do not edit by hand
\name{write_nml}
\alias{write_nml}
\title{write GLM *.nml for a GLM simulation}
\usage{
write_nml(glm_nml, file = "../resources/glm_test.nml")
}
\arguments{
\item{glm_nml}{a nml (a list) for GLM config}

\item{file}{a string with the path to the glm.nml file to write}
}
\description{
Creates a *.nml file running a GLM simulation.  \cr
}
\examples{
nml_file <- system.file('extdata', 'glm.nml', package = 'rGLM')
glm_nml <- read_nml(nml_file)
write_nml(glm_nml, file= file)
}
\author{
Jordan S. Read
}
\seealso{
\link{get_nml_value}, \link{read_nml}, \link{pretty_nml}
}
\keyword{methods}

