% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_nml.R
\name{write_nml}
\alias{write_nml}
\title{write GLM .nml for a GLM simulation}
\usage{
write_nml(glm_nml, file)
}
\arguments{
\item{glm_nml}{a nml (a list) for GLM config}

\item{file}{a string with the path to the glm2.nml file to write}
}
\description{
Creates a .nml file running a GLM simulation.  \cr
}
\examples{
glm_nml <- read_nml()
write_path <- paste0(tempdir(),'glm2.nml')
write_nml(glm_nml, file = write_path)
print(read_nml(write_path))
}
\seealso{
\link{get_nml_value}, \link{read_nml}
}
\author{
Jordan S. Read
}
\keyword{methods}
