% Generated by roxygen2 (4.0.2): do not edit by hand
\name{write_nml}
\alias{write_nml}
\title{write GLM .nml for a GLM simulation}
\usage{
write_nml(glm_nml, file)
}
\arguments{
\item{glm_nml}{a nml (a list) for GLM config}

\item{file}{a string with the path to the glm.nml file to write}
}
\description{
Creates a .nml file running a GLM simulation.  \cr
}
\examples{
glm_nml <- read_nml()
write_nml(glm_nml, file = '../glm_test.nml')
}
\author{
Jordan S. Read
}
\seealso{
\link{get_nml_value}, \link{read_nml}
}
\keyword{methods}

